+++ src/pages/checkin/checkin.scss
@@ -1,3 +1,3 @@
-page-about {
+page-checkin {
 
 }
