+++ src/pages/estabelecimentos/estabelecimentos.html
@@ -1,0 +1,27 @@
+<ion-header>
+  <ion-navbar color=primary>
+    <ion-title>Cardappio</ion-title>
+  </ion-navbar>
+  <ion-toolbar color=primary>
+    <ion-segment [(ngModel)]="estabelecimentos" color=light>
+      <ion-segment-button value="mapa">
+        <ion-icon name="navigate"></ion-icon>
+      </ion-segment-button>
+      <ion-segment-button value="lista">
+        <ion-icon name="list"></ion-icon>
+      </ion-segment-button>
+    </ion-segment>
+  </ion-toolbar>
+</ion-header>
+
+<div [ngSwitch]="estabelecimentos">
+
+  <div *ngSwitchCase="'mapa'">
+    <ion-nav [root]="mapaView" #content></ion-nav>
+  </div>
+
+  <div *ngSwitchCase="'lista'">
+    <ion-nav [root]="listaView" #content></ion-nav>
+  </div>
+
+</div>
