+++ src/pages/estabelecimentos/estabelecimentos.scss
@@ -1,0 +1,11 @@
+page-estabelecimentos {
+    ion-header {
+        position: relative;
+    }
+
+    .map-search {
+        position: absolute;
+        top: 0;
+        width: 100%;
+    }
+}
