+++ src/pages/estabelecimentos/lista.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+import { NavController } from 'ionic-angular';
+
+import { DataService } from '../../services/data-service';
+
+@Component({
+  selector: 'view-lista',
+  templateUrl: 'lista.html'
+})
+export class ListaView {
+
+}
