+++ src/pages/home/home.html
@@ -1,0 +1,9 @@
+<ion-header>
+  <ion-navbar color=primary>
+    <ion-title>Início</ion-title>
+  </ion-navbar>
+</ion-header>
+
+<ion-content padding>
+
+</ion-content>
