+++ src/pages/home/home.ts
@@ -1,0 +1,14 @@
+import { Component } from '@angular/core';
+import { NavController, AlertController } from 'ionic-angular';
+
+@Component({
+  selector: 'page-home',
+  templateUrl: 'home.html'
+})
+export class HomePage {
+
+  constructor(public navCtrl: NavController, public alertCtrl: AlertController) {
+
+  }
+
+}
