--- src/pages/mapa/mapa.scss
@@ -1,7 +1,0 @@
-page-mapa {
-    #map{
-        height: 100%;
-        width: 100%;
-    }
-
-}
