+++ src/pages/tabs/tabs.html
@@ -1,5 +1,5 @@
 <ion-tabs color=primary>
-  <ion-tab [root]="tab1Root" tabTitle="Mapa" tabIcon="navigate"></ion-tab>
+  <ion-tab [root]="tab1Root" tabTitle="Início" tabIcon="home"></ion-tab>
   <ion-tab [root]="tab2Root" tabTitle="Check-In" tabIcon="qr-scanner"></ion-tab>
-  <ion-tab [root]="tab3Root" tabTitle="Favoritos" tabIcon="heart"></ion-tab>
+  <ion-tab [root]="tab3Root" tabTitle="Estabelecimentos" tabIcon="restaurant"></ion-tab>
 </ion-tabs>
